requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
pypdf
ollamafreeapi
python-multipart
uvicorn